app/src/main/java/com/id/zul/submission5kade/presenter/match/PreviousMatchPresenter.kt
package com.id.zul.submission5kade.presenter.match

import com.google.gson.Gson
import com.id.zul.submission5kade.api.Get
import com.id.zul.submission5kade.api.Request
import com.id.zul.submission5kade.coroutine.ProviderContext
import com.id.zul.submission5kade.model.match.Previous
import com.id.zul.submission5kade.view.match.PreviousMatchView
import kotlinx.coroutines.GlobalScope
import kotlinx.coroutines.launch

class PreviousMatchPresenter(
    private val view: PreviousMatchView,
    private val apiRepository: Request,
    private val gson: Gson,
    private val context: ProviderContext = ProviderContext()
) {

    fun getPreviousMatch(query: String) {
        view.setLoading()
        GlobalScope.launch(context.main) {
            val data = gson.fromJson(
                apiRepository.getRequestAsync(Get.getPreviousMatch(query)).await(),
                Previous::class.java
            )
            view.setInItData(data.previous)
            view.unSetLoading()
        }
    }

}